requirements.txt
aiosqlite==0.22.1
anyio==4.13.0
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.7
typing_extensions==4.15.0
utils==1.0.2
